LRAC-12806 remove feature flag

+++ modules/apps/analytics/analytics-batch-export-import-api/src/main/java/com/liferay/analytics/batch/exportimport/model/listener/BaseAnalyticsDXPEntityModelListener.java
@@ -20,7 +20,6 @@ import com.liferay.analytics.settings.configuration.AnalyticsConfiguration;
 import com.liferay.analytics.settings.configuration.AnalyticsConfigurationRegistry;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.ModelListenerException;
-import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.BaseModel;
@@ -58,6 +57,10 @@ public abstract class BaseAnalyticsDXPEntityModelListener
 
 	@Override
 	public void onAfterRemove(T model) throws ModelListenerException {
+		if (!analyticsConfigurationRegistry.isActive()) {
+			return;
+		}
+
 		ShardedModel shardedModel = (ShardedModel)model;
 
 		analyticsAssociationLocalService.deleteAnalyticsAssociations(
@@ -77,9 +80,7 @@ public abstract class BaseAnalyticsDXPEntityModelListener
 
 	@Override
 	public void onBeforeRemove(T model) throws ModelListenerException {
-		if (!FeatureFlagManagerUtil.isEnabled("LRAC-10632") ||
-			!isTracked(model)) {
-
+		if (!analyticsConfigurationRegistry.isActive() || !isTracked(model)) {
 			return;
 		}
 
@@ -190,9 +191,7 @@ public abstract class BaseAnalyticsDXPEntityModelListener
 		String associationClassName, Object associationClassPK,
 		Object classPK) {
 
-		if (!FeatureFlagManagerUtil.isEnabled("LRAC-10632") ||
-			!analyticsConfigurationRegistry.isActive()) {
-
+		if (!analyticsConfigurationRegistry.isActive()) {
 			return;
 		}
